refactor(core): split ConversionEvent.kt into one file per type
Move ConversionPhase, FileResult, RunConfig, and RunStats into their
own files following the single responsibility principle.
